% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MDS_MAD.R
\name{plot_MAD_MDS}
\alias{plot_MAD_MDS}
\title{MDS and MAD Plots for Variability Analysis}
\usage{
plot_MAD_MDS(cpm, metadata, filter)
}
\arguments{
\item{cpm}{Matrix. CPM expression matrix with genes in rows and samples in columns.}

\item{metadata}{Data frame. Must contain the columns `SampleType` and `Batch`, and row names
must correspond to the samples (identifiers matching the columns of the cpm matrix).}

\item{filter}{Numeric. Threshold to filter extreme MAD values in the density plots.}
}
\value{
A list containing:
\describe{
  \item{mds_result}{Matrix with the first two dimensions of the MDS analysis}
  \item{plot_mds_group}{MDS plot by sample type}
  \item{plot_mds_batch}{MDS plot by batch}
  \item{plot_mad_group}{MAD distribution by sample type}
  \item{plot_mad_batch}{MAD distribution by batch}
}
}
\description{
This function generates plots to analyze variability in gene expression data normalized as CPM.
It performs a multidimensional scaling (MDS) analysis to visualize the similarity between samples
according to their sample type and batch. It also calculates the median absolute deviation (MAD)
of gene expression by sample type and by batch.
}
