% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_batch_effect.R
\name{correct_batch_effect}
\alias{correct_batch_effect}
\title{Batch effect correction in gene expression data using technical replicates}
\usage{
correct_batch_effect(ref_batch_n, counts, metadata, replicates)
}
\arguments{
\item{ref_batch_n}{Integer. Batch number used as the reference batch for normalization.}

\item{counts}{Matrix. Raw count data rounded to integers with samples in columns and genes in rows.}

\item{metadata}{Data frame. With two columns: \code{SampleType} (sample classification: control, disease, replicate, etc)
and \code{Batch} (batch number). Rownames must correspond to the samples and match with the colnames from \code{counts}.}

\item{replicates}{Character vector. Labels in \code{SampleType} to indicate replicate samples.}
}
\value{
A list with:
\describe{
  \item{counts_corrected}{Matrix count corrected}
  \item{adjustment_report}{Correction factors applied to each gene per batch}
  \item{cpm}{Original matrix in CPM}
  \item{cpm_corrected}{Corrected matrix in CPM}
}
}
\description{
This function performs batch effect correction on count data by using clinical replicates across different batches.
The correction is based on the relative dispersion between replicates, adjusting the counts accordingly by gene.
}
